+++ CarrierWave/CarrierWaveApp.swift
@@ -56,7 +56,8 @@ struct CarrierWaveApp: App {
         ])
         let modelConfiguration = ModelConfiguration(
             schema: schema,
-            isStoredInMemoryOnly: false
+            isStoredInMemoryOnly: false,
+            cloudKitDatabase: .none
         )
 
         do {